Handle test/unit loading minitest in some cases.
The whole minitest and test/unit loading is a bit complex. It depends
greatly on which Ruby version is being used, and if either the
'test_unit' or 'minitest' gems are available.

The newer Ruby versions (1.9+) use test/unit as a light shim around
minitest. Due to this, they load minitest when test/unit is loaded. This
means we should add the new `assertions` accessors when `:test_unit` is
specified.

However, if either Ruby 1.8 is used or the 'test_unit' gem was loaded,
'minitest' will not be loaded. In these situations we should not be
adding the accessor.